+++ Estimation/BamProcessing/BamController.cpp
@@ -14,12 +14,13 @@ namespace Estimation
 namespace BamProcessing
 {
 	BamController::BamController(const BamTags &tags, bool filled_bam, const std::string &read_param_filenames,
-	                             const std::string &gtf_path, bool gene_in_chromosome_name)
+	                             const std::string &gtf_path, bool gene_in_chromosome_name, int min_barcode_quality)
 		: _tags(tags)
 		, _filled_bam(filled_bam)
 		, _gene_in_chromosome_name(gene_in_chromosome_name)
 		, _read_param_filenames(read_param_filenames)
 		, _gtf_path(gtf_path)
+		, _min_barcode_quality(min_barcode_quality)
 	{}
 
 	void BamController::parse_bam_files(const std::vector<std::string> &bam_files, bool print_result_bams,
@@ -29,7 +30,7 @@ namespace BamProcessing
 		Tools::trace_time("Start parse bams");
 
 		auto processor = std::shared_ptr<BamProcessorAbstract>(new BamProcessor(container, this->_tags, print_result_bams));
-		BamController::process_bam_files(bam_files, processor);
+		this->process_bam_files(bam_files, processor);
 
 		Tools::trace_time("Bams parsed");
 	}
@@ -41,7 +42,7 @@ namespace BamProcessing
 		Tools::trace_time("Start write filtered bam");
 
 		auto processor = std::shared_ptr<BamProcessorAbstract>(new FilteringBamProcessor(this->_tags, container));
-		BamController::process_bam_files(bam_files, processor);
+		this->process_bam_files(bam_files, processor);
 
 		Tools::trace_time("Filtered bam written");
 	}
@@ -61,7 +62,7 @@ namespace BamProcessing
 
 		for (size_t i = 0; i < bam_files.size(); ++i)
 		{
-			BamController::parse_bam_file(bam_files[i], processor, parser, bam_files.size() == 1);
+			this->parse_bam_file(bam_files[i], processor, parser, bam_files.size() == 1);
 			processor->trace_state(bam_files[i]);
 		}
 	}
@@ -103,7 +104,7 @@ namespace BamProcessing
 				processor->trace_state(bam_name);
 			}
 
-			BamController::process_alignment(parser, processor, unexpected_chromosomes, chr_name, alignment);
+			this->process_alignment(parser, processor, unexpected_chromosomes, chr_name, alignment);
 		}
 
 		reader.Close();
@@ -130,6 +131,12 @@ namespace BamProcessing
 		if (!parser->get_read_params(alignment, read_params))
 		{
 			processor->inc_cant_parse_num();
+			return;
+		}
+
+		if (!read_params.check_quality(this->_min_barcode_quality))
+		{
+			processor->inc_low_quality_num();
 			return;
 		}