+++ Estimation/BamProcessing/BamTags.h
@@ -10,9 +10,11 @@ namespace Estimation
 		class BamTags
 		{
 		public:
-			const std::string cb;
+			const std::string cell_barcode;
 			const std::string umi;
 			const std::string gene;
+			const std::string cb_quality;
+			const std::string umi_quality;
 
 			BamTags();
 			BamTags(const boost::property_tree::ptree &config);